js folder restructure, add boilerplate
