+++ javascript/test/javascript_deps/boilerplate.js
@@ -1,0 +1,7 @@
+/**
+ * @param {number} n
+ * @return {string} src path
+ */
+var src_path = (n) => `../src/code_${n}`
+
+module.exports = { src_path };
