+++ javascript/test/test_2620.js
@@ -3,9 +3,10 @@ var chaiAsPromised = require("chai-as-promised");
 chai.use(chaiAsPromised);
 var expect = chai.expect;
 
-const { createCounter } = require("../../src/js/code_2620.js");
+var { src_path } = require("./javascript_deps/boilerplate")
+const { createCounter } = require(src_path(2620));
 
-counter = createCounter(10);
+counter = createCounter(1);
 expect(counter()).to.equal(10);
 expect(counter()).to.equal(11);
 expect(counter()).to.equal(12);
