+++ javascript/test/test_2626.js
@@ -3,7 +3,8 @@ var chaiAsPromised = require("chai-as-promised");
 chai.use(chaiAsPromised);
 var expect = chai.expect;
 
-const { reduce } = require("../../src/js/code_2626.js");
+var { src_path } = require("./javascript_deps/boilerplate")
+const { reduce } = require(src_path(2626));
 
 expect(
   reduce(
