+++ javascript/test/test_2634.js
@@ -3,7 +3,8 @@ var chaiAsPromised = require("chai-as-promised");
 chai.use(chaiAsPromised);
 var expect = chai.expect;
 
-const { filter } = require("../../src/js/code_2634.js");
+var { src_path } = require("./javascript_deps/boilerplate")
+const { filter } = require(src_path(2634));
 
 expect(
   filter(
