+++ javascript/test/test_2667.js
@@ -3,7 +3,8 @@ var chaiAsPromised = require("chai-as-promised");
 chai.use(chaiAsPromised);
 var expect = chai.expect;
 
-const { createHelloWorld } = require("../../src/js/code_2667.js");
+var { src_path } = require("./javascript_deps/boilerplate")
+const { createHelloWorld } = require(src_path(2667));
 
 const f = createHelloWorld();
 expect(f()).to.equal("Hello World");
