+++ python/test/test_65.py
@@ -1,5 +1,5 @@
 from python.src.code_65 import Solution
-from python.tests.python_deps.boilerplate import BoilerTest
+from python.test.python_deps.boilerplate import BoilerTest
 
 
 def input_fetcher(case):
